+++ crates/runner-host/src/board.rs
@@ -65,6 +65,7 @@ impl Api for Board {
     type Debug = debug::Impl;
     type Led = led::Impl;
     type Platform = Unsupported;
+    type Radio = Unsupported;
     type Rng = rng::Impl;
     type Storage = storage::Impl;
     type Timer = timer::Impl;